Métodp de abate e o de produto por parcela

+++ MinhaAPI/Domain/Services/ProdutoService.cs
@@ -9,10 +9,13 @@ namespace MinhaAPI.Domain.Services
     public class ProdutoService
     {
         private readonly ProdutoRepository _produtoRepository;
+        private readonly CompraProdutoRepository _compraProdutoRepository;
 
-        public ProdutoService(ProdutoRepository produtoRepository)
+        public ProdutoService(ProdutoRepository produtoRepository,
+                              CompraProdutoRepository compraProduto)
         {
             _produtoRepository = produtoRepository;
+            _compraProdutoRepository = compraProduto;
         }
 
         public async Task<List<ProdutoModel>> ListProdutos()
@@ -40,15 +43,22 @@ namespace MinhaAPI.Domain.Services
 
         public async Task<int> UpdateProduto(ProdutoModel produto)
         {
+            bool produtoEstaVinculadoACompra = await _compraProdutoRepository.ProdutoVinculadoACompra(produto.Id);
+            if (produtoEstaVinculadoACompra)
+                throw new InvalidOperationException("Não é possível atualizar um produto que já esteja vinculado a uma compra.");
+
             return await _produtoRepository.UpdateProduto(produto);
         }
 
         public async Task<bool> DeleteProduto(int produtoId)
         {
             ProdutoModel findProduto = await _produtoRepository.GetProduto(produtoId);
-
             if (findProduto == null)
                 throw new ArgumentException("Produto não existe!");
+
+            bool produtoEstaVinculadoACompra = await _compraProdutoRepository.ProdutoVinculadoACompra(produtoId);
+            if (produtoEstaVinculadoACompra)
+                throw new InvalidOperationException("Não é possível apagar um produto que já esteja vinculado a uma compra.");
 
             await _produtoRepository.DeleteProduto(produtoId);
